# docs/swagger.yaml
basePath: /v0
definitions:
...
  inputs.NewContactProjectAlertGroup:
    properties:
      call:
        default: false
        type: boolean
      contact_id:
        type: integer
      email:
        default: false
        type: boolean
      severity:
        type: integer
      sms:
        default: false
        type: boolean
      webhook:
        default: false
        type: boolean
    required:
    - call
    - contact_id
    - email
    - severity
    - sms
    - webhook
    type: object
...
  inputs.UpdateContactProjectAlertGroup:
    properties:
      call:
        default: false
        type: boolean
      email:
        default: false
        type: boolean
      severity:
        type: integer
      sms:
        default: false
        type: boolean
      webhook:
        default: false
        type: boolean
    type: object
  models.Response:
    properties:
      data:
        type: object
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: I have no specific description
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hod Hod API
  version: "1.0"
paths:
...
  /alert-group/{id}/project/{pid}/contact:
    get:
      description: return ContactProjectsAlertGroup
      operationId: get_contacts_of_ProjectAlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group_id
        in: path
        name: pid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get ContactProjectsAlertGroup
      tags:
      - Contact Project Alert Group
    post:
      description: new ContactProjectALertGroup
      operationId: new_ContactProjectALertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group_id
        in: path
        name: pid
        required: true
        type: integer
      - description: contact_project_alert_group
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/inputs.NewContactProjectAlertGroup'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: new ContactProjectALertGroup
      tags:
      - Contact Project Alert Group
  /alert-group/{id}/project/{pid}/contact/{cpid}:
    delete:
      description: delete ContactProjectAlertGroup by ContactProjectAlertGroupID
      operationId: delete_contactProjectAlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group_id
        in: path
        name: pid
        required: true
        type: integer
      - description: contact_project_alert_group_id
        in: path
        name: cpid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: delete ContactProjectAlertGroup
      tags:
      - Contact Project Alert Group
    get:
      description: return ContactprojectAlertGroup
      operationId: get_ContactprojectAlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group_id
        in: path
        name: pid
        required: true
        type: integer
      - description: contact_project_alert_group_id
        in: path
        name: cpid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get ContactprojectAlertGroup
      tags:
      - Contact Project Alert Group
    put:
      description: update ContactProjectAlertGroup
      operationId: update_ContactProjectAlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group_id
        in: path
        name: pid
        required: true
        type: integer
      - description: contact_project_alert_group_id
        in: path
        name: cpid
        required: true
        type: integer
      - description: contact_project_alert_group
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/inputs.UpdateContactProjectAlertGroup'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: update ContactProjectAlertGroup
      tags:
      - Contact Project Alert Group
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"